pyproject.toml
[project]
name = "random-pycon-2024-bot"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "notebook>=7.2.1",
    "litestar[full]>=2.9.1",
    "python-telegram-bot[all]>=21.3",
    "pydantic-settings>=2.3.4",
    "granian>=1.5.1",
    "aiosqlite>=0.20.0",
]
readme = "README.md"
requires-python = ">= 3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.5.3",
    "uv>=0.2.25",
    "mypy>=1.10.1",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/random_pycon_2024_bot"]

[tool.mypy]
python_version = "3.11"
plugins = [
  "pydantic.mypy"
]

follow_imports = "silent"
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true
disallow_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.rye.scripts]
devserver = "uvicorn src.random_pycon_2024_bot.app:app --reload --host 0.0.0.0"
server = "granian --interface asgi src.random_pycon_2024_bot.app:app --reload --host 0.0.0.0"
